Babylon authoritarian general changes (#101)
* Added Secession, IG Approval Setup

* Changed Approval Multipliers for lawgroups

* Change IG approvals min/max

* Changed Legitimacy

* Modified Radicals/Loyalists from Legitimacy

to keep inline with the legitimacy level changes

* Removed Legitimacy Changes

Not an appropriate time to implement them

* Final Changes

for Authoritarian general changes. Balance Patch Changes are pending

* Forgot this

Authority deficit penalty

* Update babylon_defines.txt

* lowered enactment time of diplo stance

* Added original values

added for each of the new values in defines and law groups

* Updated Min and Max Approval

This will show players the exact and real amount their current approval with an IG sits at, for better or worse.